Error en Detalle de Venta

+++ springboot/springboot/src/main/java/com/joa/springboot/DetalleVentaBarra/DetalleVentaBarraService.java
@@ -14,7 +14,6 @@ public class DetalleVentaBarraService {
     private final DetalleVentaBarraRepository detalleVentaBarraRepository;
     private final ProductoRepository productoRepository;
 
-    // Constructor injection
     public DetalleVentaBarraService(DetalleVentaBarraRepository detalleVentaBarraRepository, ProductoRepository productoRepository) {
         this.detalleVentaBarraRepository = detalleVentaBarraRepository;
         this.productoRepository = productoRepository;
@@ -24,16 +23,14 @@ public class DetalleVentaBarraService {
         Producto producto = productoRepository.findById(requestDTO.getproductoId())
                 .orElseThrow(() -> new RuntimeException("Producto no encontrado con ID: " + requestDTO.getproductoId()));
 
-        // Usar el constructor parametrizado para mayor claridad
         DetalleVentaBarra detalleVentaBarra = new DetalleVentaBarra(null, producto, requestDTO.getCantidad());
-
         detalleVentaBarra = detalleVentaBarraRepository.save(detalleVentaBarra);
 
         return new DetalleVentaBarraResponseDTO(
                 detalleVentaBarra.getId(),
                 producto.getNombre(),
                 detalleVentaBarra.getCantidad(),
-                detalleVentaBarra.getSubTotal()
+                detalleVentaBarra.getSubTotal().doubleValue()
         );
     }
 
@@ -43,11 +40,11 @@ public class DetalleVentaBarraService {
                         detalle.getId(),
                         detalle.getProducto().getNombre(),
                         detalle.getCantidad(),
-                        detalle.getSubTotal()))
+                        detalle.getSubTotal().doubleValue()))
                 .collect(Collectors.toList());
     }
 
     public void deleteDetalleVentaBarra(Long id) {
-        detalleVentaBarraRepository.deleteById(id); // Elimina directamente; JPA lanzará una excepción si no existe.
+        detalleVentaBarraRepository.deleteById(id);
     }
 }
